First commit
First commit in RPS project! I should have done this when I got the
getComputerChoice to respond to me with r, p, or s - I don't want to
discuss how long that took - but I'm only a little past that.

HTML file is blank except the title and one line of text.

script is populated with a bunch of comments, some lines to get the
computer to choose r, p, or s and a prompt to get the player to reply
with their choice. That choice is not being stored or compared to the
computer yet. Those come next.