Add play() function to play RPS
Added the play() function. It works! I wasn't adding a return (see out).
Shoutout to @/north korean supermarket on the TOP Discord for the help
with that one.

To future Wren, or anyone else who's reading these commits: consider
reading all of the steps before throwing yourself in. You made this
project harder than it needed to be - if you had kept reading, you'd
see they hold your hand a bit more, since this is your first JS project!
You didn't even need to use a prompt yet. :facepalm:

Next steps are adding a loop to limit the number of times we can play
the game. Then this RPS game is complete (but I know we revisit it
later, to spruce it up).