Make new script.js file, move old to main.js
As it says on the label.

I was struggling to make the counter work for my loop on the last
script (I should probably mention here that I did not commit when I
wrote the loop and I should have. Okay, continuing:) so I started
reading other people's scripts on their RPS projects.

That's when I realized how much of a mess my code was! And that's fine,
this is my first JS project and it's okay to have messy code. But it
was hard to understand what I had written, and by that point, I had
messed up my code trying to /fix/ the problem.

So I decided to rename the file to main.js and start fresh in my
script file. There's a good chance I'll abandon main.js.