Initial commit for main.js (prev script.js)
As it says on the label.

This was originally known as script.js. Read the commit on that file
to see why I changed the names around. This file, unfortunately, does
not match the last commit from when it was script (I added a loop, then
started trying to make a counter work, which it didn't, so there's a
ton of mess in here now). I intend to abandon this file but I'm
attached to my first messy script, so it gets to live here. Treat her
nicely.